refactor(tests): Move PostTest.php from src/Tests/views to tests/views
